+++ docs/changelog.md
@@ -1,5 +1,18 @@
-# Nhật ký thay đổi
-## [Beta 02](#beta-02)
+# [#beta-03](Beta 0.3)
+**Bổ sung - Thay đổi**
+* Khái niệm package, bổ sung packages cms-dev cung cấp nhiều artisan cho package
+* Giao diện trang quản trị hiển thị đẹp trên trên các thiết bị màn hình lớn
+* Đưa phambinhcms lên packagist
+* Thay đổi cách cài đặt Phambinhcms
+
+**Fix lỗi**
+* Fix lỗi hiển thị không đúng ngôn ngữ
+* Fix lỗi hàm kiểm tra cập nhật trong trang module chức năng và module theme
+* Fix lỗi hiển thị tại các trang tinh tức, trang tĩnh
+* Fix lỗi không cập nhật thông tin tác giả
+
+---------------
+# [#beta-02](Beta 0.2)
 **Bổ sung - Thay đổi**
 * Giao diện cài đặt có gợi ý chi tiết, chạy cài đặt hiển thị progress bar thể hiện tiến trình cài đặt
 * Hỗ trợ cài đặt bằng terminal `php artisan cms:install`
@@ -10,4 +23,3 @@
 
 **Fix lỗi**
 * Fix lỗi hiển thị không đúng ngôn ngữ
-* Các trang lỗi 404, 403 hiển thị giao diện có css